Implementing a completely patched ops watcher
