+++ .circleci/config.yml
@@ -55,22 +55,22 @@ jobs:
           at: .
       - restore_cache:
           keys:
-            - v1-optimism-build-
+            - v2-optimism-build-
       - run:
           name: Build docker containers if necessary
           command: |
             if [ ! -d ./optimism ]; then
-              npx hardhat ops --fresh --build --build-ops --optimism-path ./optimism
+              npx hardhat ops --fresh --build --build-ops
             fi;
       - save_cache:
-          key: v1-optimism-build-
+          key: v2-optimism-build-
           paths:
             - ./optimism
       - run:
           name: Start docker chains
           background: true
           command: |
-            npx hardhat ops --start --optimism-path ./optimism
+            npx hardhat ops --start
       - cmd-wait-for-port:
           port: 8545
       - cmd-wait-for-port:
