+++ .circleci/src/jobs/job-integration-tests.yml
@@ -7,22 +7,22 @@ steps:
       at: .
   - restore_cache:
       keys:
-        - v1-optimism-build-{{ checksum "package-lock.json" }}
+        - v2-optimism-build-{{ checksum "package-lock.json" }}
   - run:
       name: Build docker containers if necessary
       command: |
         if [ ! -d ./optimism ]; then
-          npx hardhat ops --fresh --build --build-ops --optimism-path ./optimism
+          npx hardhat ops --fresh --build --build-ops
         fi;
   - save_cache:
-      key: v1-optimism-build-{{ checksum "package-lock.json" }}
+      key: v2-optimism-build-{{ checksum "package-lock.json" }}
       paths:
         - ./optimism
   - run:
       name: Start docker chains
       background: true
       command: |
-        npx hardhat ops --start --optimism-path ./optimism
+        npx hardhat ops --start
   - cmd-wait-for-port:
       port: 8545
   - cmd-wait-for-port:
