+++ hardhat/tasks/task-ops.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const { homedir } = require('os');
-const { gray, yellow } = require('chalk');
+const { gray, yellow, red } = require('chalk');
 const { task } = require('hardhat/config');
 const { spawn } = require('child_process');
 const execa = require('execa');
@@ -20,22 +20,23 @@ task('ops', 'Run Optimism chain')
 	.addFlag('start', 'Start the latest build')
 	.addFlag('stop', 'Stop optimism chain')
 	.addFlag('detached', 'Detach the chain from the console')
-	.addOptionalParam('optimismPath', 'Path to optmism repository folder', '~/optimism')
+	.addOptionalParam('optimismPath', 'Path to optmism repository folder', './optimism')
+	.addOptionalParam('optimismBranch', 'Branch to checkout', 'master')
 	.addOptionalParam(
 		'optimismCommit',
 		'Commit to checkout',
-		undefined // Specify a specific commit in the branch to pin the tool
+		'd9fd67d2502a590e116ffdb6c1c53003a045e318'
 	)
 	.setAction(async (taskArguments, hre, runSuper) => {
 		taskArguments.maxMemory = true;
 
-		const branch = 'regenesis/0.4.0';
-
 		const opsPath = taskArguments.optimismPath.replace('~', homedir);
+		const opsBranch = 'regenesis/0.4.0';
 		const opsCommit = taskArguments.optimismCommit;
 		const opsDetached = taskArguments.detached ? '-d' : '';
 
-		console.log(gray('optimism commit:', taskArguments.optimismCommit));
+		console.log(gray('optimism branch:', opsBranch));
+		console.log(gray('optimism commit:', opsCommit));
 		console.log(gray('optimism folder:', opsPath));
 
 		if (taskArguments.stop) {
@@ -60,14 +61,14 @@ task('ops', 'Run Optimism chain')
 				_fresh({ opsPath });
 			}
 
-			_build({ opsPath, opsCommit, branch });
+			_build({ opsPath, opsCommit, opsBranch });
 		}
 
 		if (taskArguments.buildOps || (taskArguments.fresh && taskArguments.start)) {
 			console.log(yellow('building ops'));
 			if (!fs.existsSync(opsPath)) {
 				_fresh({ opsPath });
-				_build({ opsPath, opsCommit, branch });
+				_build({ opsPath, opsCommit, opsBranch });
 			}
 			_buildOps({ opsPath });
 		}
@@ -81,10 +82,10 @@ task('ops', 'Run Optimism chain')
 
 			if (!fs.existsSync(opsPath)) {
 				_fresh({ opsPath });
-				_build({ opsPath, opsCommit, branch });
+				_build({ opsPath, opsCommit, opsBranch });
 				_buildOps({ opsPath });
 			} else if (!_imagesExist()) {
-				_build({ opsPath, opsCommit, branch });
+				_build({ opsPath, opsCommit, opsBranch });
 				_buildOps({ opsPath });
 			}
 			await _start({ opsPath, opsDetached });
@@ -140,11 +141,11 @@ function _fresh({ opsPath }) {
 	]);
 }
 
-function _build({ opsPath, opsCommit, branch }) {
+function _build({ opsPath, opsCommit, opsBranch }) {
 	console.log(gray('  checkout commit:', opsCommit));
 	execa.sync('sh', ['-c', `cd ${opsPath} && git fetch `]);
-	execa.sync('sh', ['-c', `cd ${opsPath} && git checkout ${branch} `]);
-	execa.sync('sh', ['-c', `cd ${opsPath} && git pull origin ${branch} `]);
+	execa.sync('sh', ['-c', `cd ${opsPath} && git checkout ${opsBranch} `]);
+	execa.sync('sh', ['-c', `cd ${opsPath} && git pull origin ${opsBranch} `]);
 	if (opsCommit) {
 		execa.sync('sh', ['-c', `cd ${opsPath} && git checkout ${opsCommit}`]);
 	}
@@ -174,3 +175,12 @@ function _stop({ opsPath }) {
 	console.log(gray('  stop ops'));
 	execa.sync('sh', ['-c', `cd ${opsPath}/ops && docker-compose down -v`]);
 }
+
+function restartRelayer({ opsPath }) {
+	console.log(red('Restarting relayer subprocess...'));
+	execa.sync('sh', ['-c', `cd ${opsPath}/ops && docker-compose restart relayer`]);
+}
+
+module.exports = {
+	restartRelayer,
+};
